legacy/euro-2024-score-predictor/requirements.txt
streamlit
pandas
joblib
Pillow
scikit-learn==1.3.0
